#212 add media file repository and set cover art image from URL
